퀴즈 전체적으로 조정, 채팅 오버플로우 막음
