+++ src/main/java/com/finalproject/team4/shouldbe/service/QuizService.java
@@ -1,15 +1,51 @@
 package com.finalproject.team4.shouldbe.service;
 
+import com.finalproject.team4.shouldbe.mapper.QuizMapper;
 import com.finalproject.team4.shouldbe.vo.QuizVO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface QuizService {
-    QuizVO selectRandomQuiz(String userId, int level);
+import java.util.List;
+import java.util.Random;
 
-    String getCorrectAnswers(int quizId);
+@Service
+public class QuizService{
 
-    boolean saveQuiz(int quizId, String userId);
+    @Autowired
+    private QuizMapper QuizMapper;
 
-    void createQuiz(QuizVO quizVO);
+    public QuizVO selectRandomQuiz(String userId, int level) {
+        List<QuizVO> quizzes = QuizMapper.findQuiz(level, userId);
+        if (quizzes.isEmpty()) {
+            return null;
+        }
+        return quizzes.get(new Random().nextInt(quizzes.size()));
+    }
 
-    void saveQuizAnswer(int quizId, String userId, String correct);
+    public boolean saveQuiz(int quizId, String userId) {
+        boolean exists = QuizMapper.existsQuizInfo(quizId, userId);
+        if (!exists) {
+            QuizMapper.insertQuizInfo(quizId, userId);
+            return true;
+        }
+        return false;
+    }
+
+    public void saveQuizAnswer(int quizId, String userId, String correct) {
+        if (correct.equals("correct")) {
+            QuizMapper.insertUserAnswerInfo(quizId, userId);
+        }
+    }
+
+    public String getCorrectAnswers(int quizId) {
+        List<String> answers = QuizMapper.findAnswersByQuizId(quizId);
+        return String.join(", ", answers);
+    }
+
+    @Transactional
+    public void createQuiz(QuizVO quizVO) {
+        QuizMapper.insertQuiz(quizVO);
+        QuizMapper.insertAnswer(quizVO);
+    }
 }
