+++ src/main/resources/mapper/quiz_mapper.xml
@@ -6,6 +6,11 @@
         SELECT *
         FROM quiz
         WHERE level = #{level}
+          AND quiz_id NOT IN (
+            SELECT quiz_id
+            FROM user_quiz
+            WHERE user_id = #{userId}
+        )
     </select>
 
     <select id="findAnswersByQuizId" resultType="String">
@@ -17,25 +22,18 @@
     <select id="existsQuizInfo" resultType="boolean">
         SELECT CASE WHEN COUNT(*) > 0 THEN true ELSE false END
         FROM save_quiz
-        WHERE quiz_id = #{quiz_id}
-          AND user_id = #{user_id}
-    </select>
-
-    <select id="existsUserAnswerInfo" resultType="boolean">
-        SELECT CASE WHEN COUNT(*) > 0 THEN true ELSE false END
-        FROM user_quiz
-        WHERE quiz_id = #{quiz_id}
-          AND user_id = #{user_id}
+        WHERE quiz_id = #{quizId}
+          AND user_id = #{userId}
     </select>
 
     <insert id="insertQuizInfo">
         INSERT INTO save_quiz (quiz_id, user_id)
-        VALUES (#{quiz_id}, #{user_id})
+        VALUES (#{quizId}, #{userId})
     </insert>
 
     <insert id="insertUserAnswerInfo">
         INSERT INTO user_quiz (quiz_id, user_id)
-        VALUES (#{quiz_id}, #{user_id})
+        VALUES (#{quizId}, #{userId})
     </insert>
 
     <insert id="insertQuiz" useGeneratedKeys="true" keyProperty="quiz_id">
